Support dumping Seqs of values
I want to be able to dump a Seq of values that I can then process in the C test
harness.  This patch dumps each element of the array as a seperate #define,
along with a "__COUNT" at the end.  This format is amenable to processing with
C preprocessor repat macros.